added more tables to db at create_table.sql and switched LOGIN from Domain to Service
