+++ project/DBQueries/dropTables.sql
@@ -1,0 +1,51 @@
+-- A VERY DANGEROUS FILE - WATCH OUT!!!
+
+IF OBJECT_ID('dbo.AssociationRepresentative', 'U') IS NOT NULL
+DROP TABLE dbo.AssociationRepresentative
+GO
+
+IF OBJECT_ID('dbo.Admins', 'U') IS NOT NULL
+DROP TABLE dbo.Admins
+GO
+
+
+
+IF OBJECT_ID('dbo.UsersFavoriteGames', 'U') IS NOT NULL
+DROP TABLE dbo.UsersFavoriteGames
+GO
+
+IF OBJECT_ID('dbo.Events', 'U') IS NOT NULL
+DROP TABLE dbo.Events
+GO
+
+IF OBJECT_ID('dbo.Users', 'U') IS NOT NULL
+DROP TABLE dbo.Users
+GO
+
+IF OBJECT_ID('dbo.SeasonReferees', 'U') IS NOT NULL
+DROP TABLE dbo.SeasonReferees
+GO
+
+IF OBJECT_ID('dbo.Referees', 'U') IS NOT NULL
+DROP TABLE dbo.Referees
+GO
+
+IF OBJECT_ID('dbo.Games', 'U') IS NOT NULL
+DROP TABLE dbo.Games
+GO
+IF OBJECT_ID('dbo.GamePolicy', 'U') IS NOT NULL
+DROP TABLE dbo.GamePolicy
+GO
+
+IF OBJECT_ID('dbo.Season', 'U') IS NOT NULL
+DROP TABLE dbo.Season
+GO
+
+
+IF OBJECT_ID('dbo.League', 'U') IS NOT NULL
+DROP TABLE dbo.League
+GO
+
+
+
+
