+++ project/Domain/Player.js
@@ -1,0 +1,18 @@
+
+/*player object {
+    username: string,
+    password: string,
+    name: name,
+    isConnected: boolean,
+    qualification: string,
+    birthDate: date
+}
+*/
+
+function addPersonalContent(content){
+    //adds content
+}
+
+function editPersonalInformation(edit){
+    //edit
+}
